pyproject.toml
[tool.poetry]
name = "ssl_expiry_monitor"
version = "0.1.0"
description = "Lambda Function (Script) to Report on hosts SSL expiry"
authors = ["johnpreston <john@compose-x.io>"]
license = "MPL-2.0"
classifiers = [
  "Natural Language :: English",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.8"
compose-x-common = "^1.2.12"
pyopenssl = "^23.1.1"

[tool.poetry.group.dev.dependencies]
cleanpy = "^0.4.0"
pyclean = "^2.7.0"
black = "^23.3"
isort = "^5.9.3"
coverage = "^7.2"
pytest = "^7.2"
pre-commit = "^3.3"
tbump = "^6.9"
pyupgrade = "^3.4.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "kelvin"

[tool.coverage.report]
exclude_lines = [
  "if __name__ == '__main__'"
]
[tool.coverage.run]
omit = [
  "*/cli.py"
]

[tool.tbump]
github_url = "https://github.com/johnpreston/ssl_expiry_monitor"

[tool.tbump.version]
current = "0.1.0"
regex = '''
  (?P<major>\d+)
  \.
  (?P<minor>\d+)
  \.
  (?P<patch>\d+)
  (?:(?P<rc>[\S]+))?
  '''


[tool.tbump.git]
message_template = "Bump to {new_version}"
tag_template = "v{new_version}"

[[tool.tbump.file]]
src = "pyproject.toml"

[[tool.tbump.file]]
src = "ssl_expiry_monitor/__init__.py"

[[tool.tbump.before_commit]]
name = "Files format"
cmd = "make conform"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
